+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
     Failed to parse "IpAddress": Invalid IP address: abc (occurred while
     parsing "[IpAddress!]")
     ```
+- Added the `theme` field to the `Account` struct to store the user's
+  selected screen color mode. Accordingly, the functions for inserting and
+  updating accounts have been modified, and new APIs have been added to retrieve
+  and update the user's selected screen color mode.
 
 ### Changed